Make suggested changes incl responsivenes;  profile popper toggle
